refactor(cluster): split distributor prepare and runtime tests
