+++ src/agilab/core/test/test_agi_distributor.py
@@ -574,152 +574,6 @@ def test_worker_site_packages_dir_and_pth_writer_branches(tmp_path):
     assert pth_path.read_text(encoding="utf-8").endswith("_uv_sources\n")
 
 
-def test_ensure_asyncio_run_signature_patches_pydevd_shim(monkeypatch):
-    original = agi_distributor_module.asyncio.run
-
-    def _fake_run(main, debug=None):
-        return ("orig", main, debug)
-
-    _fake_run.__module__ = "pydevd.fake"
-    monkeypatch.setattr(agi_distributor_module.asyncio, "run", _fake_run)
-    agi_distributor_module._ensure_asyncio_run_signature()
-
-    patched = agi_distributor_module.asyncio.run
-    assert patched is not _fake_run
-    assert patched("task", debug=True) == ("orig", "task", True)
-
-    async def _coro():
-        return 7
-
-    assert patched(_coro(), loop_factory=asyncio.new_event_loop) == 7
-    monkeypatch.setattr(agi_distributor_module.asyncio, "run", original)
-
-
-def test_ensure_asyncio_run_signature_tolerates_event_loop_runtime_errors(monkeypatch):
-    original = agi_distributor_module.asyncio.run
-
-    def _fake_run(main, debug=None):
-        return ("orig", main, debug)
-
-    _fake_run.__module__ = "pydevd.fake"
-    monkeypatch.setattr(agi_distributor_module.asyncio, "run", _fake_run)
-
-    set_calls: list[object] = []
-
-    def _fake_set_event_loop(loop):
-        set_calls.append(loop)
-        raise RuntimeError("loop policy locked")
-
-    monkeypatch.setattr(agi_distributor_module.asyncio, "set_event_loop", _fake_set_event_loop)
-    agi_distributor_module._ensure_asyncio_run_signature()
-
-    class _Loop:
-        def __init__(self):
-            self.debug = None
-            self.closed = False
-            self.awaited = []
-
-        def set_debug(self, value):
-            self.debug = value
-
-        def run_until_complete(self, main):
-            self.awaited.append(main)
-            return "done"
-
-        def close(self):
-            self.closed = True
-
-    loop = _Loop()
-    patched = agi_distributor_module.asyncio.run
-    assert patched("task", debug=True, loop_factory=lambda: loop) == "done"
-    assert loop.debug is True
-    assert loop.closed is True
-    assert set_calls == [loop, None]
-    monkeypatch.setattr(agi_distributor_module.asyncio, "run", original)
-
-
-def test_ensure_asyncio_run_signature_leaves_non_pydevd_shim_untouched(monkeypatch):
-    def _fake_run(main, debug=None):
-        return ("orig", main, debug)
-
-    _fake_run.__module__ = "custom.runner"
-    monkeypatch.setattr(agi_distributor_module.asyncio, "run", _fake_run)
-    agi_distributor_module._ensure_asyncio_run_signature()
-    assert agi_distributor_module.asyncio.run is _fake_run
-
-
-def test_ensure_asyncio_run_signature_handles_uninspectable_run(monkeypatch):
-    def _fake_run(main, debug=None):
-        return ("orig", main, debug)
-
-    monkeypatch.setattr(agi_distributor_module.asyncio, "run", _fake_run)
-    monkeypatch.setattr(
-        agi_distributor_module.inspect,
-        "signature",
-        lambda *_a, **_k: (_ for _ in ()).throw(ValueError("no signature")),
-    )
-    agi_distributor_module._ensure_asyncio_run_signature()
-    assert agi_distributor_module.asyncio.run is _fake_run
-
-
-def test_agi_version_missing_on_pypi_detection(tmp_path, monkeypatch):
-    project = tmp_path / "project"
-    project.mkdir(parents=True, exist_ok=True)
-
-    # no pyproject
-    assert agi_distributor_module._agi__version_missing_on_pypi(project) is False
-
-    pyproject = project / "pyproject.toml"
-    pyproject.write_text("[project]\nname='demo'\n", encoding="utf-8")
-    assert agi_distributor_module._agi__version_missing_on_pypi(project) is False
-
-    pyproject.write_text('agi-core = "==1.2.3"\n', encoding="utf-8")
-
-    class _Resp:
-        def __init__(self, payload):
-            self._payload = payload
-
-        def __enter__(self):
-            return io.StringIO(json.dumps(self._payload))
-
-        def __exit__(self, exc_type, exc, tb):
-            return False
-
-    monkeypatch.setattr(
-        urllib.request,
-        "urlopen",
-        lambda *_a, **_k: _Resp({"releases": {"1.2.3": [{}]}}),
-    )
-    assert agi_distributor_module._agi__version_missing_on_pypi(project) is False
-
-    monkeypatch.setattr(
-        urllib.request,
-        "urlopen",
-        lambda *_a, **_k: _Resp({"releases": {"1.2.4": [{}]}}),
-    )
-    assert agi_distributor_module._agi__version_missing_on_pypi(project) is True
-
-    monkeypatch.setattr(
-        urllib.request,
-        "urlopen",
-        lambda *_a, **_k: (_ for _ in ()).throw(RuntimeError("network down")),
-    )
-    assert agi_distributor_module._agi__version_missing_on_pypi(project) is False
-
-
-def test_format_exception_chain_compacts_causes():
-    try:
-        try:
-            raise ValueError("inner")
-        except ValueError as exc:
-            raise RuntimeError("RuntimeError: inner") from exc
-    except Exception as exc:
-        text = agi_distributor_module._format_exception_chain(exc)
-
-    assert "inner" in text
-    assert ("RuntimeError" in text) or ("ValueError" in text)
-
-
 def test_background_process_manager_tracks_running_completed_and_dead_jobs(monkeypatch, tmp_path):
     popen_calls: list[dict[str, object]] = []
 
@@ -898,15 +752,6 @@ async def test_calibration_fallback_uses_worker_counts_when_no_worker_keys_exist
 
     assert AGI._capacity == {"10.0.0.1:0": 1.0, "10.0.0.1:1": 1.0}
     assert AGI.workers_info == {"10.0.0.1:0": {"label": 1.0}, "10.0.0.1:1": {"label": 1.0}}
-
-
-def test_format_exception_chain_strips_generic_error_prefixes():
-    class CustomError(Exception):
-        pass
-
-    text = agi_distributor_module._format_exception_chain(CustomError("CustomError: precise detail"))
-
-    assert text.endswith("CustomError: precise detail")
 
 
 @pytest.fixture(autouse=True)
@@ -2769,697 +2614,6 @@ async def test_agi_service_real_dask_e2e_self_heal_submit_stop(monkeypatch, tmp_
 
 
 @pytest.mark.asyncio
-async def test_prepare_local_env_offline_worker_env_initializes_project(monkeypatch, tmp_path):
-    wenv_abs = tmp_path / "wenv_local"
-    AGI.env = SimpleNamespace(
-        wenv_abs=wenv_abs,
-        python_version="3.13",
-        verbose=1,
-        envars={"AGI_INTERNET_ON": "0"},
-        uv="uv",
-        is_worker_env=True,
-        hw_rapids_capable=None,
-    )
-    AGI._rapids_enabled = True
-    captured_set_env = []
-    run_calls = []
-
-    async def _fake_detect(_ip):
-        return "export PATH=\"$HOME/.local/bin:$PATH\"; "
-
-    async def _fake_run(cmd, cwd):
-        run_calls.append((cmd, str(cwd)))
-        return ""
-
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "_hardware_supports_rapids", staticmethod(lambda: False))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run", staticmethod(_fake_run))
-    monkeypatch.setattr(
-        agi_distributor_module.AgiEnv,
-        "set_env_var",
-        staticmethod(lambda *args: captured_set_env.append(args)),
-    )
-    monkeypatch.setattr(agi_distributor_module.distributor_cli, "python_version", lambda: "3.13.12")
-
-    await AGI._prepare_local_env()
-
-    assert wenv_abs.exists()
-    assert any(entry[0] == "127.0.0.1_CMD_PREFIX" for entry in captured_set_env)
-    assert any(entry[0] == "127.0.0.1_PYTHON_VERSION" for entry in captured_set_env)
-    assert any("no_rapids_hw" in entry for entry in captured_set_env)
-    assert len(run_calls) == 1
-    assert "--project" in run_calls[0][0]
-    assert "init --bare --no-workspace" in run_calls[0][0]
-
-
-@pytest.mark.asyncio
-async def test_prepare_local_env_online_handles_python_download_warning(monkeypatch, tmp_path):
-    wenv_abs = tmp_path / "wenv_local"
-    AGI.env = SimpleNamespace(
-        wenv_abs=wenv_abs,
-        python_version="3.13",
-        verbose=1,
-        envars={"AGI_INTERNET_ON": "1"},
-        uv="uv",
-        is_worker_env=False,
-        hw_rapids_capable=None,
-    )
-    AGI._rapids_enabled = True
-    run_calls = []
-
-    async def _fake_detect(_ip):
-        return ""
-
-    async def _fake_run(cmd, _cwd):
-        run_calls.append(cmd)
-        if "python install" in cmd:
-            raise RuntimeError("No download found for request")
-        return ""
-
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "_hardware_supports_rapids", staticmethod(lambda: True))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run", staticmethod(_fake_run))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-    monkeypatch.setattr(agi_distributor_module.distributor_cli, "python_version", lambda: "3.13.12")
-    monkeypatch.setattr(agi_distributor_module.os, "name", "posix", raising=False)
-
-    await AGI._prepare_local_env()
-
-    assert any("self update" in cmd for cmd in run_calls)
-    assert any("python install 3.13" in cmd for cmd in run_calls)
-
-
-@pytest.mark.asyncio
-async def test_prepare_local_env_online_ignores_uv_self_update_failure(monkeypatch, tmp_path):
-    wenv_abs = tmp_path / "wenv_local"
-    AGI.env = SimpleNamespace(
-        wenv_abs=wenv_abs,
-        python_version="3.13",
-        verbose=1,
-        envars={"AGI_INTERNET_ON": "1"},
-        uv="uv",
-        is_worker_env=False,
-        hw_rapids_capable=None,
-    )
-    AGI._rapids_enabled = True
-    run_calls = []
-
-    async def _fake_detect(_ip):
-        return ""
-
-    async def _fake_run(cmd, _cwd):
-        run_calls.append(cmd)
-        if "self update" in cmd:
-            raise RuntimeError("Self-update is only available for standalone installs")
-        return ""
-
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "_hardware_supports_rapids", staticmethod(lambda: True))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run", staticmethod(_fake_run))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-    monkeypatch.setattr(agi_distributor_module.distributor_cli, "python_version", lambda: "3.13.12")
-    monkeypatch.setattr(agi_distributor_module.os, "name", "posix", raising=False)
-
-    await AGI._prepare_local_env()
-
-    assert any("self update" in cmd for cmd in run_calls)
-    assert any("python install 3.13" in cmd for cmd in run_calls)
-
-
-@pytest.mark.asyncio
-async def test_prepare_local_env_windows_skips_self_update_when_standalone_uv_missing(monkeypatch, tmp_path):
-    wenv_abs = tmp_path / "wenv_local"
-    fake_home = tmp_path / "home"
-    fake_home.mkdir(parents=True, exist_ok=True)
-
-    AGI.env = SimpleNamespace(
-        wenv_abs=wenv_abs,
-        python_version="3.13",
-        verbose=1,
-        envars={"AGI_INTERNET_ON": "1"},
-        uv="uv --quiet",
-        is_worker_env=False,
-        hw_rapids_capable=None,
-    )
-    AGI._rapids_enabled = True
-    run_calls = []
-    warnings = []
-
-    async def _fake_detect(_ip):
-        return "set PATH=%USERPROFILE%\\\\.local\\\\bin;%PATH% && "
-
-    async def _fake_run(cmd, _cwd):
-        run_calls.append(cmd)
-        return ""
-
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "_hardware_supports_rapids", staticmethod(lambda: True))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run", staticmethod(_fake_run))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-    monkeypatch.setattr(agi_distributor_module.distributor_cli, "python_version", lambda: "3.13.12")
-    monkeypatch.setattr(agi_distributor_module.os, "name", "nt", raising=False)
-    monkeypatch.setattr(agi_distributor_module.Path, "home", classmethod(lambda cls: fake_home))
-    monkeypatch.setattr(agi_distributor_module.logger, "warning", lambda *args, **kwargs: warnings.append(args))
-
-    await AGI._prepare_local_env()
-
-    standalone_uv = fake_home / ".local" / "bin" / "uv.exe"
-    assert not any("self update" in cmd for cmd in run_calls)
-    assert any("python install 3.13" in cmd for cmd in run_calls)
-    assert any(str(standalone_uv) in str(entry[1]) for entry in warnings if len(entry) > 1)
-
-
-@pytest.mark.asyncio
-async def test_prepare_local_env_windows_uses_standalone_uv_when_available(monkeypatch, tmp_path):
-    wenv_abs = tmp_path / "wenv_local"
-    fake_home = tmp_path / "home"
-    standalone_uv = fake_home / ".local" / "bin" / "uv.exe"
-    standalone_uv.parent.mkdir(parents=True, exist_ok=True)
-    standalone_uv.write_text("", encoding="utf-8")
-
-    AGI.env = SimpleNamespace(
-        wenv_abs=wenv_abs,
-        python_version="3.13",
-        verbose=1,
-        envars={"AGI_INTERNET_ON": "1"},
-        uv="uv --quiet",
-        is_worker_env=False,
-        hw_rapids_capable=None,
-    )
-    AGI._rapids_enabled = True
-    run_calls = []
-
-    async def _fake_detect(_ip):
-        return "set PATH=%USERPROFILE%\\\\.local\\\\bin;%PATH% && "
-
-    async def _fake_run(cmd, _cwd):
-        run_calls.append(cmd)
-        return ""
-
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "_hardware_supports_rapids", staticmethod(lambda: True))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run", staticmethod(_fake_run))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-    monkeypatch.setattr(agi_distributor_module.distributor_cli, "python_version", lambda: "3.13.12")
-    monkeypatch.setattr(agi_distributor_module.os, "name", "nt", raising=False)
-    monkeypatch.setattr(agi_distributor_module.Path, "home", classmethod(lambda cls: fake_home))
-
-    await AGI._prepare_local_env()
-
-    assert any(str(standalone_uv) in cmd and "self update" in cmd for cmd in run_calls)
-    assert any("uv --quiet python install 3.13" in cmd for cmd in run_calls)
-
-
-@pytest.mark.asyncio
-async def test_prepare_cluster_env_happy_path_sends_files(monkeypatch, tmp_path):
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-
-    worker_pyproject = tmp_path / "worker_pyproject.toml"
-    worker_pyproject.write_text("[project]\nname='demo-worker'\n", encoding="utf-8")
-    manager_pyproject = tmp_path / "manager_pyproject.toml"
-    manager_pyproject.write_text("[project]\nname='demo-manager'\n", encoding="utf-8")
-    uvproject = tmp_path / "uv.toml"
-    uvproject.write_text("[tool.uv]\n", encoding="utf-8")
-
-    env = SimpleNamespace(
-        dist_rel=Path("wenv/dist"),
-        wenv_rel=Path("wenv"),
-        pyvers_worker="3.13",
-        is_local=lambda ip: ip == "127.0.0.1",
-        envars={"AGI_INTERNET_ON": "1"},
-        uv="uv",
-        cluster_pck=cluster_pck,
-        worker_pyproject=worker_pyproject,
-        manager_pyproject=manager_pyproject,
-        uvproject=uvproject,
-        target_worker="demo_worker",
-    )
-    AGI.env = env
-    AGI._workers = {"10.0.0.2": 1}
-    AGI.agi_workers = {"pandas": "pandas-worker"}
-    sent = []
-    remote_cmds = []
-
-    async def _fake_detect(_ip):
-        return "export PATH=\"$HOME/.local/bin:$PATH\"; "
-
-    async def _fake_exec(ip, cmd):
-        remote_cmds.append((ip, cmd))
-        if "--version" in cmd:
-            return "uv 0.6.0"
-        return "ok"
-
-    async def _fake_send(_env, ip, files, remote_path, user=None, password=None):
-        sent.append(
-            (
-                ip,
-                [{"name": Path(f).name, "content": Path(f).read_text(encoding="utf-8")} for f in files],
-                str(remote_path),
-            )
-        )
-
-    async def _fake_noop(*_args, **_kwargs):
-        return None
-
-    monkeypatch.setattr(AGI, "_get_scheduler", staticmethod(lambda _scheduler: ("127.0.0.1", 8786)))
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_exec))
-    monkeypatch.setattr(AGI, "send_files", staticmethod(_fake_send))
-    monkeypatch.setattr(AGI, "_kill", staticmethod(_fake_noop))
-    monkeypatch.setattr(AGI, "_clean_dirs", staticmethod(_fake_noop))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-
-    await AGI._prepare_cluster_env("127.0.0.1")
-
-    assert any("self update" in cmd for _, cmd in remote_cmds)
-    assert any(item[2] == "wenv" for item in sent)
-    assert any(any(item["name"] == "cli.py" for item in items) for _, items, _ in sent)
-
-
-@pytest.mark.asyncio
-async def test_prepare_cluster_env_stages_uv_source_payload(monkeypatch, tmp_path):
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-
-    deps_root = tmp_path / "deps" / "ilp_worker"
-    deps_root.mkdir(parents=True, exist_ok=True)
-    (deps_root / "pyproject.toml").write_text("[project]\nname='ilp_worker'\n", encoding="utf-8")
-    (deps_root / "milp.py").write_text("class MILP: pass\n", encoding="utf-8")
-
-    worker_dir = tmp_path / "worker_src"
-    worker_dir.mkdir(parents=True, exist_ok=True)
-    worker_pyproject = worker_dir / "pyproject.toml"
-    worker_pyproject.write_text(
-        """
-[project]
-name='demo-worker'
-[tool.uv.sources]
-ilp_worker = { path = "../deps/ilp_worker" }
-""".strip(),
-        encoding="utf-8",
-    )
-    manager_pyproject = tmp_path / "manager_pyproject.toml"
-    manager_pyproject.write_text("[project]\nname='demo-manager'\n", encoding="utf-8")
-    uvproject = tmp_path / "uv.toml"
-    uvproject.write_text("[tool.uv]\n", encoding="utf-8")
-
-    env = SimpleNamespace(
-        dist_rel=Path("wenv/dist"),
-        wenv_rel=Path("wenv"),
-        pyvers_worker="3.13",
-        is_local=lambda ip: ip == "127.0.0.1",
-        envars={"AGI_INTERNET_ON": "1"},
-        uv="uv",
-        cluster_pck=cluster_pck,
-        worker_pyproject=worker_pyproject,
-        manager_pyproject=manager_pyproject,
-        uvproject=uvproject,
-        target_worker="demo_worker",
-        verbose=1,
-    )
-    AGI.env = env
-    AGI._workers = {"10.0.0.2": 1}
-    AGI.agi_workers = {"dag": "dag-worker"}
-    sent = []
-
-    async def _fake_detect(_ip):
-        return "export PATH=\"$HOME/.local/bin:$PATH\"; "
-
-    async def _fake_exec(_ip, cmd):
-        if "--version" in cmd:
-            return "uv 0.6.0"
-        return "ok"
-
-    async def _fake_send(_env, ip, files, remote_path, user=None, password=None):
-        items = []
-        for f in files:
-            p = Path(f)
-            item = {"name": p.name, "is_dir": p.is_dir()}
-            if p.is_file():
-                item["content"] = p.read_text(encoding="utf-8")
-            elif p.is_dir():
-                item["children"] = sorted(child.relative_to(p).as_posix() for child in p.rglob("*"))
-            items.append(item)
-        sent.append((ip, items, str(remote_path)))
-
-    async def _fake_noop(*_args, **_kwargs):
-        return None
-
-    monkeypatch.setattr(AGI, "_get_scheduler", staticmethod(lambda _scheduler: ("127.0.0.1", 8786)))
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_exec))
-    monkeypatch.setattr(AGI, "send_files", staticmethod(_fake_send))
-    monkeypatch.setattr(AGI, "_kill", staticmethod(_fake_noop))
-    monkeypatch.setattr(AGI, "_clean_dirs", staticmethod(_fake_noop))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-
-    await AGI._prepare_cluster_env("127.0.0.1")
-
-    sent_to_wenv = [items for _ip, items, remote_path in sent if remote_path == "wenv"]
-    assert sent_to_wenv
-    payload = sent_to_wenv[0]
-    pyproject_item = next(item for item in payload if item["name"] == "pyproject.toml")
-    assert '_uv_sources/ilp_worker' in pyproject_item["content"]
-    staged_item = next(item for item in payload if item["name"] == "_uv_sources")
-    assert staged_item["is_dir"] is True
-    assert "ilp_worker/milp.py" in staged_item["children"]
-
-
-@pytest.mark.asyncio
-async def test_prepare_cluster_env_ignores_uv_self_update_failure(monkeypatch, tmp_path):
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-
-    worker_pyproject = tmp_path / "worker_pyproject.toml"
-    worker_pyproject.write_text("[project]\nname='demo-worker'\n", encoding="utf-8")
-    manager_pyproject = tmp_path / "manager_pyproject.toml"
-    manager_pyproject.write_text("[project]\nname='demo-manager'\n", encoding="utf-8")
-    uvproject = tmp_path / "uv.toml"
-    uvproject.write_text("[tool.uv]\n", encoding="utf-8")
-
-    env = SimpleNamespace(
-        dist_rel=Path("wenv/dist"),
-        wenv_rel=Path("wenv"),
-        pyvers_worker="3.13",
-        is_local=lambda ip: ip == "127.0.0.1",
-        envars={"AGI_INTERNET_ON": "1"},
-        uv="uv",
-        cluster_pck=cluster_pck,
-        worker_pyproject=worker_pyproject,
-        manager_pyproject=manager_pyproject,
-        uvproject=uvproject,
-        target_worker="demo_worker",
-    )
-    AGI.env = env
-    AGI._workers = {"10.0.0.2": 1}
-    AGI.agi_workers = {"pandas": "pandas-worker"}
-    sent = []
-    remote_cmds = []
-
-    async def _fake_detect(_ip):
-        return "export PATH=\"$HOME/.local/bin:$PATH\"; "
-
-    async def _fake_exec(ip, cmd):
-        remote_cmds.append((ip, cmd))
-        if "--version" in cmd:
-            return "uv 0.6.0"
-        if "self update" in cmd:
-            raise RuntimeError("Self-update is only available for standalone installs")
-        return "ok"
-
-    async def _fake_send(_env, ip, files, remote_path, user=None, password=None):
-        sent.append(
-            (
-                ip,
-                [{"name": Path(f).name, "content": Path(f).read_text(encoding="utf-8")} for f in files],
-                str(remote_path),
-            )
-        )
-
-    async def _fake_noop(*_args, **_kwargs):
-        return None
-
-    monkeypatch.setattr(AGI, "_get_scheduler", staticmethod(lambda _scheduler: ("127.0.0.1", 8786)))
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_exec))
-    monkeypatch.setattr(AGI, "send_files", staticmethod(_fake_send))
-    monkeypatch.setattr(AGI, "_kill", staticmethod(_fake_noop))
-    monkeypatch.setattr(AGI, "_clean_dirs", staticmethod(_fake_noop))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-
-    await AGI._prepare_cluster_env("127.0.0.1")
-
-    assert any("self update" in cmd for _, cmd in remote_cmds)
-    assert any("python install 3.13" in cmd for _, cmd in remote_cmds)
-    assert any(item[2] == "wenv" for item in sent)
-
-
-@pytest.mark.asyncio
-async def test_prepare_cluster_env_uses_manager_pyproject_when_worker_missing(monkeypatch, tmp_path):
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-
-    manager_pyproject = tmp_path / "manager_pyproject.toml"
-    manager_pyproject.write_text("[project]\nname='demo-manager'\n", encoding="utf-8")
-    worker_pyproject = tmp_path / "missing_worker.toml"
-    uvproject = tmp_path / "uv.toml"
-    uvproject.write_text("[tool.uv]\n", encoding="utf-8")
-
-    env = SimpleNamespace(
-        dist_rel=Path("wenv/dist"),
-        wenv_rel=Path("wenv"),
-        pyvers_worker="3.13",
-        is_local=lambda ip: ip == "127.0.0.1",
-        envars={"AGI_INTERNET_ON": "1"},
-        uv="uv",
-        cluster_pck=cluster_pck,
-        worker_pyproject=worker_pyproject,
-        manager_pyproject=manager_pyproject,
-        uvproject=uvproject,
-        target_worker="demo_worker",
-    )
-    AGI.env = env
-    AGI._workers = {"10.0.0.2": 1}
-    AGI.agi_workers = {"pandas": "pandas-worker"}
-    sent = []
-
-    async def _fake_detect(_ip):
-        return "export PATH=\"$HOME/.local/bin:$PATH\"; "
-
-    async def _fake_exec(_ip, cmd):
-        if "--version" in cmd:
-            return "uv 0.6.0"
-        return "ok"
-
-    async def _fake_send(_env, ip, files, remote_path, user=None, password=None):
-        sent.append(
-            (
-                ip,
-                [{"name": Path(f).name, "content": Path(f).read_text(encoding="utf-8")} for f in files],
-                str(remote_path),
-            )
-        )
-
-    async def _fake_noop(*_args, **_kwargs):
-        return None
-
-    monkeypatch.setattr(AGI, "_get_scheduler", staticmethod(lambda _scheduler: ("127.0.0.1", 8786)))
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_exec))
-    monkeypatch.setattr(AGI, "send_files", staticmethod(_fake_send))
-    monkeypatch.setattr(AGI, "_kill", staticmethod(_fake_noop))
-    monkeypatch.setattr(AGI, "_clean_dirs", staticmethod(_fake_noop))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-
-    await AGI._prepare_cluster_env("127.0.0.1")
-
-    sent_to_wenv = [items for _ip, items, remote_path in sent if remote_path == "wenv"]
-    assert sent_to_wenv
-    assert any(
-        any(item["name"] == "pyproject.toml" and "demo-manager" in item["content"] for item in items)
-        for items in sent_to_wenv
-    )
-    assert any(any(item["name"] == "uv.toml" for item in items) for items in sent_to_wenv)
-
-
-@pytest.mark.asyncio
-async def test_prepare_cluster_env_uses_unique_stage_dir_and_cleans_it(monkeypatch, tmp_path):
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-
-    worker_pyproject = tmp_path / "worker_pyproject.toml"
-    worker_pyproject.write_text("[project]\nname='demo-worker'\n", encoding="utf-8")
-    manager_pyproject = tmp_path / "manager_pyproject.toml"
-    manager_pyproject.write_text("[project]\nname='demo-manager'\n", encoding="utf-8")
-    uvproject = tmp_path / "uv.toml"
-    uvproject.write_text("[tool.uv]\n", encoding="utf-8")
-
-    env = SimpleNamespace(
-        dist_rel=Path("wenv/dist"),
-        wenv_rel=Path("wenv"),
-        pyvers_worker="3.13",
-        is_local=lambda ip: ip == "127.0.0.1",
-        envars={"AGI_INTERNET_ON": "1"},
-        uv="uv",
-        cluster_pck=cluster_pck,
-        worker_pyproject=worker_pyproject,
-        manager_pyproject=manager_pyproject,
-        uvproject=uvproject,
-        target_worker="demo_worker",
-    )
-    AGI.env = env
-    AGI._workers = {"10.0.0.2": 1}
-    AGI.agi_workers = {"pandas": "pandas-worker"}
-    sent = []
-    staged_roots = []
-
-    async def _fake_detect(_ip):
-        return "export PATH=\"$HOME/.local/bin:$PATH\"; "
-
-    async def _fake_exec(_ip, cmd):
-        if "--version" in cmd:
-            return "uv 0.6.0"
-        return "ok"
-
-    async def _fake_send(_env, ip, files, remote_path, user=None, password=None):
-        sent.append(
-            (
-                ip,
-                [(Path(f).name, Path(f).exists()) for f in files],
-                str(remote_path),
-            )
-        )
-
-    async def _fake_noop(*_args, **_kwargs):
-        return None
-
-    def _fake_mkdtemp(prefix):
-        stage_root = tmp_path / f"{prefix}stage"
-        stage_root.mkdir(parents=True, exist_ok=False)
-        staged_roots.append(stage_root)
-        return str(stage_root)
-
-    def _fake_stage_sources(src_pyproject, dest_pyproject, stage_root, log_rewrites):
-        assert src_pyproject == worker_pyproject
-        assert dest_pyproject == stage_root / "pyproject.toml"
-        assert stage_root == staged_roots[0]
-        assert log_rewrites is False
-        staged_source = stage_root / "staged-source.txt"
-        staged_source.write_text("staged", encoding="utf-8")
-        return [staged_source]
-
-    monkeypatch.setattr(AGI, "_get_scheduler", staticmethod(lambda _scheduler: ("127.0.0.1", 8786)))
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_exec))
-    monkeypatch.setattr(AGI, "send_files", staticmethod(_fake_send))
-    monkeypatch.setattr(AGI, "_kill", staticmethod(_fake_noop))
-    monkeypatch.setattr(AGI, "_clean_dirs", staticmethod(_fake_noop))
-    monkeypatch.setattr(agi_distributor_module, "mkdtemp", _fake_mkdtemp)
-    monkeypatch.setattr(
-        agi_distributor_module,
-        "_stage_uv_sources_for_copied_pyproject",
-        _fake_stage_sources,
-    )
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-
-    await AGI._prepare_cluster_env("127.0.0.1")
-
-    assert staged_roots
-    assert staged_roots[0].name.startswith("agilab_demo_worker_pyproject_")
-    assert staged_roots[0].exists() is False
-    sent_to_wenv = [items for _ip, items, remote_path in sent if remote_path == "wenv"]
-    assert sent_to_wenv
-    assert any(name == "pyproject.toml" and exists for name, exists in sent_to_wenv[0])
-    assert any(name == "staged-source.txt" and exists for name, exists in sent_to_wenv[0])
-    assert any(name == "uv.toml" and exists for name, exists in sent_to_wenv[0])
-
-
-@pytest.mark.asyncio
-async def test_prepare_cluster_env_falls_back_to_original_pyproject_when_extra_seed_fails(monkeypatch, tmp_path):
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-
-    worker_pyproject = tmp_path / "worker_pyproject.toml"
-    worker_pyproject.write_text("[project]\nname='demo-worker'\n", encoding="utf-8")
-    manager_pyproject = tmp_path / "manager_pyproject.toml"
-    manager_pyproject.write_text("[project]\nname='demo-manager'\n", encoding="utf-8")
-    uvproject = tmp_path / "uv.toml"
-    uvproject.write_text("[tool.uv]\n", encoding="utf-8")
-
-    env = SimpleNamespace(
-        dist_rel=Path("wenv/dist"),
-        wenv_rel=Path("wenv"),
-        pyvers_worker="3.13",
-        is_local=lambda ip: ip == "127.0.0.1",
-        envars={"AGI_INTERNET_ON": "1"},
-        uv="uv",
-        cluster_pck=cluster_pck,
-        worker_pyproject=worker_pyproject,
-        manager_pyproject=manager_pyproject,
-        uvproject=uvproject,
-        target_worker="demo_worker",
-    )
-    AGI.env = env
-    AGI._workers = {"10.0.0.2": 1}
-    AGI.agi_workers = {"pandas": "pandas-worker"}
-    sent = []
-
-    async def _fake_detect(_ip):
-        return "export PATH=\"$HOME/.local/bin:$PATH\"; "
-
-    async def _fake_exec(_ip, cmd):
-        if "--version" in cmd:
-            return "uv 0.6.0"
-        return "ok"
-
-    async def _fake_send(_env, ip, files, remote_path, user=None, password=None):
-        sent.append((ip, [Path(f).name for f in files], str(remote_path)))
-
-    async def _fake_noop(*_args, **_kwargs):
-        return None
-
-    monkeypatch.setattr(AGI, "_get_scheduler", staticmethod(lambda _scheduler: ("127.0.0.1", 8786)))
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_exec))
-    monkeypatch.setattr(AGI, "send_files", staticmethod(_fake_send))
-    monkeypatch.setattr(AGI, "_kill", staticmethod(_fake_noop))
-    monkeypatch.setattr(AGI, "_clean_dirs", staticmethod(_fake_noop))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-    monkeypatch.setattr(
-        agi_distributor_module,
-        "_ensure_optional_extras",
-        lambda *_args, **_kwargs: (_ for _ in ()).throw(RuntimeError("boom")),
-    )
-
-    await AGI._prepare_cluster_env("127.0.0.1")
-
-    sent_to_wenv = [names for _ip, names, remote_path in sent if remote_path == "wenv"]
-    assert sent_to_wenv
-    assert any("worker_pyproject.toml" in names for names in sent_to_wenv)
-    assert not any("pyproject.toml" in names and "worker_pyproject.toml" not in names for names in sent_to_wenv)
-
-
-@pytest.mark.asyncio
-async def test_prepare_cluster_env_raises_when_uv_missing_offline(monkeypatch):
-    env = SimpleNamespace(
-        dist_rel=Path("wenv/dist"),
-        wenv_rel=Path("wenv"),
-        pyvers_worker="3.13",
-        is_local=lambda ip: False,
-        envars={"AGI_INTERNET_ON": "0"},
-        uv="uv",
-        cluster_pck=Path("."),
-        worker_pyproject=Path("worker_pyproject.toml"),
-        manager_pyproject=Path("manager_pyproject.toml"),
-        uvproject=Path("uv.toml"),
-        target_worker="demo_worker",
-    )
-    AGI.env = env
-    AGI._workers = {"10.0.0.2": 1}
-
-    async def _fake_exec(_ip, _cmd):
-        raise RuntimeError("uv missing")
-
-    async def _fake_detect(_ip):
-        return ""
-
-    monkeypatch.setattr(AGI, "_get_scheduler", staticmethod(lambda _scheduler: ("127.0.0.1", 8786)))
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_exec))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "set_env_var", staticmethod(lambda *_a, **_k: None))
-
-    with pytest.raises(EnvironmentError, match="Uv binary is not installed"):
-        await AGI._prepare_cluster_env("127.0.0.1")
-
-
-@pytest.mark.asyncio
 async def test_start_scheduler_local_switches_port_and_connects(monkeypatch, tmp_path):
     cluster_pck = tmp_path / "cluster"
     (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
@@ -4420,36 +3574,6 @@ async def test_build_lib_local_selects_fireducks_package(monkeypatch, tmp_path):
 
 
 @pytest.mark.asyncio
-async def test_uninstall_modules_and_venv_todo_cover_cleanup_and_logging(monkeypatch):
-    commands = []
-    logger = mock.Mock()
-    AGI.env = SimpleNamespace(uv="uv", agi_env=Path("/tmp/agi-env"), verbose=1, logger=logger)
-    AGI._module_to_clean = ["demo-one", "demo-two"]
-
-    async def _fake_run(cmd, cwd):
-        commands.append((cmd, cwd))
-        return ""
-
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run", staticmethod(_fake_run))
-    monkeypatch.setattr(agi_distributor_module.logger, "info", logger.info)
-    monkeypatch.setattr(AgiEnv, "is_local", staticmethod(lambda ip: ip == "127.0.0.1"))
-    monkeypatch.setattr(agi_distributor_module, "env", AGI.env, raising=False)
-
-    await AGI._uninstall_modules()
-    AGI._venv_todo({"127.0.0.1", "10.0.0.2"})
-
-    assert commands == [
-        ("uv pip uninstall demo-one -y", Path("/tmp/agi-env")),
-        ("uv pip uninstall demo-two -y", Path("/tmp/agi-env")),
-    ]
-    assert AGI._module_to_clean == []
-    assert AGI._local_ip == ["127.0.0.1"]
-    assert AGI._remote_ip == ["10.0.0.2"]
-    assert AGI._install_todo == 2
-    assert logger.info.called
-
-
-@pytest.mark.asyncio
 async def test_start_scheduler_wraps_non_runtime_error(monkeypatch, tmp_path):
     app_path = tmp_path / "app"
     app_path.mkdir()
@@ -4543,83 +3667,6 @@ def test_train_capacity_missing_and_success(tmp_path):
     model_path = tmp_path / AGI._capacity_model_file
     assert model_path.exists()
     assert hasattr(AGI._capacity_predictor, "predict")
-
-
-@pytest.mark.asyncio
-async def test_prepare_cluster_env_fallback_installer_and_no_download(monkeypatch, tmp_path):
-    cluster_pck = tmp_path / "cluster_pck"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-    worker_pyproject = tmp_path / "worker_pyproject.toml"
-    worker_pyproject.write_text("[project]\nname='worker'\n", encoding="utf-8")
-    uvproject = tmp_path / "uv.toml"
-    uvproject.write_text("[tool.uv]\n", encoding="utf-8")
-
-    env = SimpleNamespace(
-        dist_rel=Path("wenv/dist"),
-        wenv_rel=Path("wenv"),
-        pyvers_worker="3.13",
-        is_local=lambda ip: ip == "127.0.0.1",
-        envars={"AGI_INTERNET_ON": "1"},
-        uv="uv",
-        cluster_pck=cluster_pck,
-        worker_pyproject=worker_pyproject,
-        manager_pyproject=tmp_path / "manager_pyproject.toml",
-        uvproject=uvproject,
-        target_worker="demo_worker",
-    )
-    env.manager_pyproject.write_text("[project]\nname='manager'\n", encoding="utf-8")
-    AGI.env = env
-    AGI._workers = {"10.0.0.2": 1}
-    AGI.agi_workers = {"pandas": "pandas-worker"}
-    sent = []
-    cmds = []
-
-    class _FakeProcessError(Exception):
-        pass
-
-    async def _fake_detect(_ip):
-        return "export PATH=\"$HOME/.local/bin:$PATH\"; "
-
-    async def _fake_exec(ip, cmd):
-        cmds.append((ip, cmd))
-        if "--version" in cmd:
-            raise RuntimeError("uv missing")
-        if "install.ps1" in cmd:
-            raise RuntimeError("windows installer failed")
-        if "curl -LsSf https://astral.sh/uv/install.sh | sh" in cmd:
-            return "ok"
-        if "python install" in cmd:
-            raise _FakeProcessError("No download found for request")
-        return "ok"
-
-    async def _fake_send(_env, ip, files, remote_path, user=None, password=None):
-        sent.append((ip, [Path(f).name for f in files], str(remote_path)))
-
-    async def _noop(*_args, **_kwargs):
-        return None
-
-    monkeypatch.setattr(AGI, "_get_scheduler", staticmethod(lambda _scheduler: ("127.0.0.1", 8786)))
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_exec))
-    monkeypatch.setattr(AGI, "send_files", staticmethod(_fake_send))
-    monkeypatch.setattr(AGI, "_kill", staticmethod(_noop))
-    monkeypatch.setattr(AGI, "_clean_dirs", staticmethod(_noop))
-    monkeypatch.setattr(agi_distributor_module, "ProcessError", _FakeProcessError)
-    monkeypatch.setattr(
-        agi_distributor_module.AgiEnv,
-        "set_env_var",
-        staticmethod(lambda key, value=None: env.envars.__setitem__(key, value)),
-    )
-
-    await AGI._prepare_cluster_env("127.0.0.1")
-
-    joined = "\n".join(cmd for _, cmd in cmds)
-    assert "install.ps1" in joined
-    assert "install.sh | sh" in joined
-    assert any("python install 3.13" in cmd for _, cmd in cmds)
-    assert any(item[2] == "wenv" for item in sent)
-    assert any("cli.py" in names for _, names, _ in sent)
 
 
 @pytest.mark.asyncio
